Splitting the answer validators to different sections - the other one was becoming way too crowded.
