+++ libs/application/templates/parental-leave/src/lib/answerValidationSections/validateLatestPeriodValidationSection.ts
@@ -1,0 +1,124 @@
+import { Application, StaticTextObject } from "@island.is/application/types"
+import { Period } from "../../types"
+import { errorMessages } from "../messages"
+import { AnswerValidationError, buildValidationError, getValueViaPath } from "@island.is/application/core"
+import isArray from "lodash/isArray"
+import { calculateDaysUsedByPeriods, filterValidPeriods, getAvailableRightsInDays, getExpectedDateOfBirth } from "../parentalLeaveUtils"
+import { validatePeriod } from "./utils"
+import { AnswerValidationConstants } from "../../constants"
+const {
+  VALIDATE_LATEST_PERIOD,
+} = AnswerValidationConstants
+
+export const validateLatestPeriodValidationSection = (newAnswer: unknown, application: Application) => {
+  let periods = newAnswer as Period[] | undefined
+
+  // If added new a period, sometime the old periods in newAnswer are 'null'
+  // If that happen, take the periods in application and use them
+  const filterPeriods = periods?.filter(
+    (period) => !!period?.startDate || !!period?.firstPeriodStart,
+  )
+
+  const periodDictionary = (periods as unknown) as Record<string, Period>
+  const invalidKey = Object.keys(periodDictionary).find((key) => {
+    const currentPeriod = periodDictionary[key]
+    return isNaN(Date.parse(currentPeriod.startDate))
+  })
+
+  if (invalidKey) {
+    return {
+      path: `periods[${invalidKey}].startDate`,
+      message: errorMessages.periodsStartDateRequired,
+    }
+  }
+
+  if (filterPeriods?.length !== periods?.length) {
+    periods = getValueViaPath(application.answers, 'periods')
+    periods = periods?.filter((period) => !!period?.startDate)
+  }
+
+  if (!isArray(periods)) {
+    return {
+      path: 'periods',
+      message: errorMessages.periodsNotAList,
+    }
+  }
+
+  if (periods?.length === 0) {
+    // Nothing to validate
+    return undefined
+  }
+
+  let daysUsedByPeriods, rights
+  try {
+    daysUsedByPeriods = calculateDaysUsedByPeriods(periods)
+    rights = getAvailableRightsInDays(application)
+  } catch (e) {
+    return {
+      path: 'periods',
+      message: e,
+      values: {},
+    }
+  }
+
+  if (daysUsedByPeriods > rights) {
+    return {
+      path: 'periods',
+      message: errorMessages.periodsExceedRights,
+      values: {
+        daysUsedByPeriods,
+        rights,
+      },
+    }
+  }
+
+  const latestPeriodIndex = periods.length - 1
+  const latestPeriod = periods[latestPeriodIndex]
+  const expectedDateOfBirth = getExpectedDateOfBirth(application)
+
+  if (!expectedDateOfBirth) {
+    return {
+      path: 'periods',
+      message: errorMessages.dateOfBirth,
+    }
+  }
+
+  const otherPeriods = periods.slice(0, latestPeriodIndex)
+  const validOtherPeriods = filterValidPeriods(otherPeriods)
+
+  const isFirstPeriod = validOtherPeriods.length === 0
+
+  const buildError = buildValidationError(
+    VALIDATE_LATEST_PERIOD,
+    latestPeriodIndex,
+  )
+
+  const buildFieldError = (
+    field: string | null,
+    message: StaticTextObject,
+    values: Record<string, unknown> = {},
+  ): AnswerValidationError => {
+    if (field !== null) {
+      return buildError(message, field, values)
+    }
+
+    return {
+      path: VALIDATE_LATEST_PERIOD,
+      message,
+      values,
+    }
+  }
+
+  const validatedField = validatePeriod(
+    latestPeriod,
+    isFirstPeriod,
+    validOtherPeriods,
+    application,
+    buildFieldError,
+  )
+
+  if (validatedField !== undefined) {
+    return validatedField
+  }
+  return undefined
+}
